Add Palindrome Checking Using Stack Code in C
This commit adds the C code for checking whether a given linked list is a palindrome or not using a stack. The code takes user input to create a linked list, inserts the elements into a stack, and compares the stack elements with the linked list in a palindrome-checking manner.

Files Added:

Palindrome_using_stack.c: This file contains the C code for checking a linked list's palindrome property using a stack. The code includes functions for inserting nodes into the linked list (insert), pushing elements into the stack (push), copying the linked list (copylist), and checking if the linked list is a palindrome or not (isPalindrome). The main function handles user input, calls the necessary functions, and displays the result.