Create palindromeCheckerADT.c
Uploading the solution for this problem - 
"Write a C program to find the palindrome using appropriate ADT to implement a more 
powerful version of the is_palindrome() function. As a reminder, this function 
implements simple palindome verification. Here is the signature and documentation for 
the function:
• bool is_palindrome(char *text)
 Return true if text is a palindrome, false otherwise. A palindrome is a string that is 
identical to itself when reversed. For example, "madam", "dad", and "abba" are 
palindromes. Note: the empty string is a palindrome, as is every string of length one.
 Your solution should ignore whitespace and punctuation, and all comparisons should be 
case-insensitive. Include some tests in your main function. Examples of valid 
palindromes:
Example- "", "a", "aa", "aaa", "aba", "abba", "Taco cat", "Madam"